+++ src/elements/Variant.php
@@ -25,6 +25,7 @@ use craft\commerce\events\CustomizeVariantSnapshotFieldsEvent;
 use craft\commerce\helpers\Purchasable as PurchasableHelper;
 use craft\commerce\models\ProductType;
 use craft\commerce\models\Sale;
+use craft\commerce\Plugin;
 use craft\commerce\records\Variant as VariantRecord;
 use craft\db\Query;
 use craft\db\Table as CraftTable;
@@ -1116,6 +1117,32 @@ class Variant extends Purchasable implements NestedElementInterface
             [['stock', 'fieldId', 'ownerId', 'primaryOwnerId'], 'number'],
             [['ownerId', 'primaryOwnerId', 'isDefault'], 'safe'],
         ]);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function availableShippingCategories(): array
+    {
+        $productTypeId = $this->getPrimaryOwner()?->getType()->id;
+        if ($productTypeId) {
+            return Plugin::getInstance()->getShippingCategories()->getShippingCategoriesByProductTypeId($productTypeId);
+        }
+
+        return parent::availableShippingCategories();
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function availableTaxCategories(): array
+    {
+        $productTypeId = $this->getPrimaryOwner()?->getType()->id;
+        if ($productTypeId) {
+            return Plugin::getInstance()->getTaxCategories()->getTaxCategoriesByProductTypeId($productTypeId);
+        }
+
+        return parent::availableTaxCategories();
     }
 
     /**