+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 ## Unreleased
 
+- Added `craft\commerce\base\Purchasable::shippingCategoryFieldHtml()`.
+- Added `craft\commerce\base\Purchasable::taxCategoryFieldHtml()`.
+- Added `craft\commerce\base\Purchasable::availableShippingCategories()`.
+- Added `craft\commerce\base\Purchasable::availableTaxCategories()`.
+- Added `craft\commerce\elements\Variant::availableShippingCategories()`.
+- Added `craft\commerce\elements\Variant::availableTaxCategories()`.
 - Fixed a SQL error that occurred when reordering order statuses on PostgreSQL. ([#3554](https://github.com/craftcms/commerce/issues/3554))
+- Fixed a bug where it was possible to select shipping and tax categories not associated to the current product type. ([#3557](https://github.com/craftcms/commerce/issues/3557))
 
 ## 5.0.11.1 - 2024-06-20